src/main/java/com/graphql/aidan/show/ShowDatafetcher.java
package com.graphql.aidan.show;

import com.netflix.graphql.dgs.DgsComponent;
import com.netflix.graphql.dgs.DgsQuery;
import com.netflix.graphql.dgs.InputArgument;

import java.util.List;
import java.util.stream.Collectors;

@DgsComponent
public class ShowDatafetcher {

    private final List<Show> shows = List.of(
            new Show("Breaking Bad", 2010),
            new Show("Game Of Thrones", 2007),
            new Show("Friends", 1990)
    );

    @DgsQuery
    public List<Show> shows(@InputArgument String titleFilter) {
        if(titleFilter == null) {
            return shows;
        }

        return shows.stream().filter(s -> s.getTitle().contains(titleFilter)).collect(Collectors.toList());
    }
}
